% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evintroPkg.R
\docType{package}
\name{evintro-package}
\alias{evintro}
\alias{evintro-package}
\title{\packageTitle{evintro}}
\description{
The DESCRIPTION file:
\packageDESCRIPTION{evintro}
\packageIndices{evintro}

The \pkg{evintro} package provides basic functions helping
at
}
