% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expplot.R
\name{expplot}
\alias{expplot}
\title{Exponential Return Level Plot}
\usage{
expplot(
  y,
  a = NULL,
  xTicks = c("T", "pExc", "Exp", "p"),
  add = FALSE,
  param = NULL,
  largestText = 0L,
  ...
)
}
\arguments{
\item{y}{A numeric vector containing block maxima or a character
defining a distribution function such as \code{"GEV"},
\code{"exp"}, ...}

\item{a}{Passed to \code{\link{ppoints}}}

\item{xTicks}{The type of ticks on the horizontal axis. With
\code{"Exp"} the ticks show the quantiles of the standard
exponential distribution.}

\item{add}{Logical. If \code{TRUE}}

\item{param}{Used only when \code{y} is character. A vector with
its named elements corresponding to the parameter arguments of
the distribution function. See the \bold{Examples} section.}

\item{largestText}{Used only when \code{y} is numeric. Number of
largest observations for which a label will be added on the
plot. The labels are extracted from the names of \code{y}. See
\bold{Examples}}

\item{...}{Arguments passed to \code{plot} or \code{points} such
as \code{main}, \code{col}, ...}
}
\value{
Nothing
}
\description{
Exponential Return Level Plot
}
\section{}{
Caution Mind that the \pkg{Renext} package has a function
    with the same name.
}

\examples{
library(evd)
y <- rexp(80)
expplot(y, xTicks = "T", pch = 16, col = "orangered", a = 0)
expplot(y, pch = 21, col = "SpringGreen3", cex = 1.2, lwd = 2,
         add = TRUE)
names(y) <- 2024 - rev(seq_along(y))
expplot(y, xTicks = "T", largestText = 4,
         pch = 16, col = "orangered", a = 0)
expplot("GPD2", param = c("shape" = -0.1),
         main = "GPD distribution on an Exponetial plot")
expplot("GPD2", param = c("shape" = 0.1), col = "red",
          add = TRUE)

}
