% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fingGEVPar.R
\name{findGEVPar}
\alias{findGEVPar}
\title{Find the GEV Parameters for a given Shape, Mean
and Quantile}
\usage{
findGEVPar(xi, mean = 0, q99 = 1)
}
\arguments{
\item{xi}{Given GEV shape.}

\item{mean}{Given GEV mean.}

\item{q99}{Given quantile with probability 0.99.}
}
\value{
A vector with its named elements being the three GEV
    parameters
}
\description{
In order to get familiar with the GEV family if can
    be useful to compare GEV densities having the same mean and
    the same quantile for the probability 0.99 because these two
    quantities can be quite well estimated from a given sample.
}
\examples{
library(nieve)
theta <- findGEVPar(xi = 0.1)
qGEV(p = 0.99, loc = theta[1], scale = theta[2], shape = theta[3])
qGEV(p = 0.999, loc = theta[1], scale = theta[2], shape = theta[3])

theta1 <- findGEVPar(xi = -0.1)
qGEV(p = 0.99, loc = theta1[1], scale = theta1[2], shape = theta1[3])
qGEV(p = 0.999, loc = theta1[1], scale = theta1[2], shape = theta1[3])
}
