% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gumbplot.R
\name{gumbplot}
\alias{gumbplot}
\title{Gumbel Return Level Plot}
\usage{
gumbplot(
  y,
  a = NULL,
  xTicks = c("T", "pExc", "Gumbel", "p"),
  add = FALSE,
  param = NULL,
  largestText = 0L,
  ...
)
}
\arguments{
\item{y}{A numeric vector containing block maxima or a character
defining a distribution function such as \code{"GEV"},
\code{"exp"}, ...}

\item{a}{Passed to \code{\link{ppoints}}}

\item{xTicks}{The type of ticks on the horizontal axis. With the
choice \code{"Gumbel"}, the ticks show the quantiles of the
standard Gumbel distribution.}

\item{add}{Logical. If \code{TRUE}}

\item{param}{Used only when \code{y} is character. A vector with
its named elements corresponding to the parameter arguments of
the distribution function. See the \bold{Examples} section.}

\item{largestText}{Used only when \code{y} is numeric. Number of
largest observations for which a label will be added on the
plot. The labels are extracted from the names of \code{y}. See
\bold{Examples}}

\item{...}{Arguments passed to \code{plot} or \code{points} such
as \code{main}, \code{col}, ...}
}
\value{
Nothing
}
\description{
Gumbel Return Level Plot
}
\examples{
library(evd)
y <- rgumbel(80)
gumbplot(y, xTicks = "T", pch = 16, col = "orangered", a = 0)
gumbplot(y, pch = 21, col = "SpringGreen3", cex = 1.2, lwd = 2,
         add = TRUE)
names(y) <- 2024 - rev(seq_along(y))
gumbplot(y, xTicks = "T", largestText = 4,
         pch = 16, col = "orangered", a = 0)
gumbplot("GEV", param = c(loc = 100, "shape" = -0.1),
         main = "GEV distribution on a Gumbel plot")
gumbplot("GEV", param = c(loc = 100, "shape" = 0.1), col = "red",
          add = TRUE)

}
